+++ packages/iov-tendermint-rpc/src/client.spec.ts
@@ -3,13 +3,14 @@ import { ReadonlyDate } from "readonly-date";
 import { Stream } from "xstream";
 
 import { Encoding } from "@iov/encoding";
+import { toListPromise } from "@iov/stream";
 
 import { Adaptor, adatorForVersion } from "./adaptor";
 import { Client } from "./client";
-import { randomId } from "./common";
+import { randomId } from "./jsonrpc";
 import { buildQuery } from "./requests";
 import * as responses from "./responses";
-import { HttpClient, RpcClient, WebsocketClient } from "./rpcclient";
+import { HttpClient, RpcClient, WebsocketClient } from "./rpcclients";
 
 function skipTests(): boolean {
   return !process.env.TENDERMINT_ENABLED;
@@ -426,9 +427,40 @@ function websocketTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor): void
     })().catch(done.fail);
   });
 
-  // This a a minimal showcase for subscribing to the same query twice
-  // See https://github.com/iov-one/iov-core/issues/581
-  xit("can subscribe twice", done => {
+  it("can unsubscribe and re-subscribe to the same stream", async () => {
+    pendingWithoutTendermint();
+
+    const client = new Client(rpcFactory(), adaptor);
+    const stream = client.subscribeNewBlockHeader();
+
+    const event1 = (await toListPromise(stream, 1))[0];
+    expect(event1.height).toBeGreaterThanOrEqual(1);
+    expect(event1.time.getTime()).toBeGreaterThanOrEqual(1);
+
+    // No sleep: producer will not be stopped in the meantime
+
+    const event2 = (await toListPromise(stream, 1))[0];
+    expect(event2.height).toBeGreaterThan(event1.height);
+    expect(event2.time.getTime()).toBeGreaterThan(event1.time.getTime());
+
+    // Very short sleep: just enough to schedule asynchonous producer stopping
+    await sleep(5);
+
+    const event3 = (await toListPromise(stream, 1))[0];
+    expect(event3.height).toBeGreaterThan(event2.height);
+    expect(event3.time.getTime()).toBeGreaterThan(event2.time.getTime());
+
+    // Proper sleep: enough to finish unsubscribing at over the network
+    await sleep(100);
+
+    const event4 = (await toListPromise(stream, 1))[0];
+    expect(event4.height).toBeGreaterThan(event3.height);
+    expect(event4.time.getTime()).toBeGreaterThan(event3.time.getTime());
+
+    client.disconnect();
+  });
+
+  it("can subscribe twice", done => {
     pendingWithoutTendermint();
 
     (async () => {