+++ packages/iov-tendermint-rpc/src/index.ts
@@ -1,5 +1,4 @@
 export { Client } from "./client";
-export { TxHash } from "./common";
 export {
   AbciInfoRequest,
   AbciQueryParams,
@@ -25,4 +24,5 @@ export {
   ValidatorsRequest,
 } from "./requests";
 export * from "./responses";
-export { HttpClient, WebsocketClient } from "./rpcclient";
+export { HttpClient, WebsocketClient } from "./rpcclients"; // TODO: Why do we export those outside of this package?
+export { IpPortString, TxHash } from "./types";