+++ packages/iov-tendermint-rpc/src/responses.ts
@@ -2,8 +2,7 @@ import { ReadonlyDate } from "readonly-date";
 
 import { Algorithm, ChainId, PostableBytes, PublicKeyBundle, SignatureBytes } from "@iov/base-types";
 
-import { TxHash } from "./common";
-import { IpPortString } from "./encodings";
+import { IpPortString, TxHash } from "./types";
 
 export type Response =
   | AbciInfoResponse