+++ packages/iov-tendermint-rpc/types/index.d.ts
@@ -1,5 +1,5 @@
 export { Client } from "./client";
-export { TxHash } from "./common";
 export { AbciInfoRequest, AbciQueryParams, AbciQueryRequest, BlockRequest, BlockchainRequest, BlockResultsRequest, BroadcastTxRequest, BroadcastTxParams, CommitRequest, GenesisRequest, HealthRequest, Method, Request, QueryString, QueryTag, StatusRequest, SubscriptionEventType, TxParams, TxRequest, TxSearchParams, TxSearchRequest, ValidatorsRequest, } from "./requests";
 export * from "./responses";
-export { HttpClient, WebsocketClient } from "./rpcclient";
+export { HttpClient, WebsocketClient } from "./rpcclients";
+export { IpPortString, TxHash } from "./types";