+++ packages/iov-tendermint-rpc/types/jsonrpc.d.ts
@@ -1,10 +1,5 @@
 import { As } from "type-tagger";
-/**
- * A raw tendermint transaction hash, currently 20 bytes
- */
-export declare type TxHash = Uint8Array & As<"tx-hash">;
 export declare type RpcId = string & As<"rpcid">;
-export declare const rpcVersion = "2.0";
 export interface JsonRpc {
     readonly jsonrpc: "2.0";
     readonly id: RpcId;
@@ -31,9 +26,10 @@ export interface JsonRpcEvent {
         readonly value: any;
     };
 }
-export declare const jsonRpc: () => JsonRpc;
-export declare const jsonRpcWith: (method: string, params?: {} | undefined) => JsonRpcRequest;
-export declare const throwIfError: (resp: JsonRpcResponse) => JsonRpcSuccess;
-export declare const ifError: (resp: JsonRpcResponse) => Error | undefined;
-export declare const randomChar: () => string;
-export declare const randomId: () => RpcId;
+export declare function jsonRpc(): JsonRpc;
+export declare function jsonRpcWith(method: string, params?: {}): JsonRpcRequest;
+export declare function throwIfError(resp: JsonRpcResponse): JsonRpcSuccess;
+export declare function ifError(resp: JsonRpcResponse): Error | undefined;
+/** generates a random alphanumeric character  */
+export declare function randomChar(): string;
+export declare function randomId(): RpcId;