+++ packages/iov-tendermint-rpc/types/rpcclients/websocketclient.d.ts
@@ -1,0 +1,19 @@
+import { Stream } from "xstream";
+import { JsonRpcEvent, JsonRpcRequest, JsonRpcResponse, JsonRpcSuccess } from "../jsonrpc";
+import { RpcStreamingClient } from "./rpcclient";
+export declare class WebsocketClient implements RpcStreamingClient {
+    private readonly bridge;
+    private readonly url;
+    private readonly socket;
+    private readonly subscriptionStreams;
+    constructor(baseUrl?: string, onError?: (err: any) => void);
+    execute(request: JsonRpcRequest): Promise<JsonRpcSuccess>;
+    listen(request: JsonRpcRequest): Stream<JsonRpcEvent>;
+    /**
+     * Resolves as soon as websocket is connected. execute() queues requests automatically,
+     * so this should be required for testing purposes only.
+     */
+    connected(): Promise<void>;
+    disconnect(): void;
+    protected responseForRequestId(id: string): Promise<JsonRpcResponse>;
+}