+++ packages/iov-tendermint-rpc/types/v0-20/index.d.ts
@@ -1,6 +1,7 @@
-import { Params } from "./requests";
-import { Responses } from "./responses";
-export declare const v0_20: {
-    params: typeof Params;
-    responses: typeof Responses;
-};
+import { Adaptor } from "../adaptor";
+/**
+ * @deprecated Support for Tendermint version 0.20.0 and 0.21.0 is deprecated
+ * and will be removed in the next version of IOV-Core. If you need support
+ * for Tendermint < 0.25.0, please open an issue on Github and we'll maintain it.
+ */
+export declare const v0_20: Adaptor;