+++ packages/iov-tendermint-rpc/types/v0-25/index.d.ts
@@ -1,6 +1,2 @@
-import { Params } from "./requests";
-import { Responses } from "./responses";
-export declare const v0_25: {
-    params: typeof Params;
-    responses: typeof Responses;
-};
+import { Adaptor } from "../adaptor";
+export declare const v0_25: Adaptor;